+++ library/FieldUtils/ProcessModules/ProcessInterpPtsToPts.cpp
@@ -200,7 +200,7 @@ void ProcessInterpPtsToPts::CreateFieldPts(po::variables_map &vm)
             }
         }
 
-        vector<int> ppe;
+        vector<size_t> ppe;
         ppe.push_back(npts);
         m_f->m_fieldPts =
             MemoryManager<LibUtilities::PtsField>::AllocateSharedPtr(dim,
@@ -270,7 +270,7 @@ void ProcessInterpPtsToPts::CreateFieldPts(po::variables_map &vm)
             }
         }
 
-        vector<int> ppe;
+        vector<size_t> ppe;
         ppe.push_back(npts[0]);
         ppe.push_back(npts[1]);
         m_f->m_fieldPts =
@@ -333,7 +333,7 @@ void ProcessInterpPtsToPts::CreateFieldPts(po::variables_map &vm)
             }
         }
 
-        vector<int> ppe;
+        vector<size_t> ppe;
         ppe.push_back(npts[0]);
         ppe.push_back(npts[1]);
         ppe.push_back(npts[2]);
