+++ library/LibUtilities/Foundations/NodalTetEvenlySpaced.h
@@ -81,7 +81,7 @@ namespace Nektar
                 const Array<OneD, const NekDouble> &y,
                 const Array<OneD, const NekDouble> &z)
             {
-                int          numpoints = x.num_elements();
+                size_t       numpoints = x.num_elements();
                 unsigned int np        = GetTotNumPoints();
 
                 Array<OneD, NekDouble> interp(GetTotNumPoints()*numpoints);
