+++ library/LibUtilities/Foundations/NodalTriEvenlySpaced.h
@@ -78,7 +78,7 @@ namespace Nektar
                 const Array<OneD, const NekDouble> &x,
                 const Array<OneD, const NekDouble> &y)
             {
-                int          numpoints = x.num_elements();
+                size_t       numpoints = x.num_elements();
                 unsigned int np        = GetTotNumPoints();
                 
                 Array<OneD, NekDouble> interp(GetTotNumPoints()*numpoints);
